changing up functions to match updated slides.

+++ parameters_arguments.py
@@ -1,4 +1,5 @@
-def add_tax(price : float, rate: float) -> float:
+def add_tax(price: int, rate: float) -> float:
     return price * (1 + rate)
-total = add_tax(10.00, 0.055)
+
+total = add_tax(10, 0.055)
 print(total)
